Move existing files out the way
