+++ keyboards/lfkeyboards/smk65/readme.md
@@ -1,14 +1,17 @@
-SMK65
-===
+# SMK65
 
 65% layout based on Mat3o's Whitefox, but with support for SMK switches.
 
-Keyboard Maintainer: [LFKeyboards](https://github.com/lfkeyboards)  
-Hardware Supported: SMK65  
-Hardware Availability: [LFKeyboards.com](https://www.lfkeyboards.com/)
+* Keyboard Maintainer: [LFKeyboards](https://github.com/lfkeyboards)
+* Hardware Supported: SMK65
+* Hardware Availability: [LFKeyboards.com](https://www.lfkeyboards.com/)
 
 Make example for this keyboard (after setting up your build environment):
 
     make lfkeyboards/smk65:default
 
-See [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) then the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information.
+Flashing example for this keyboard:
+
+    make lfkeyboards/smk65:default:flash
+
+See the [build environment setup](https://docs.qmk.fm/#/getting_started_build_tools) and the [make instructions](https://docs.qmk.fm/#/getting_started_make_guide) for more information. Brand new to QMK? Start with our [Complete Newbs Guide](https://docs.qmk.fm/#/newbs).
