+++ keyboards/lfkeyboards/smk65/revb/rules.mk
@@ -1,0 +1,15 @@
+# MCU name
+MCU = at90usb646
+
+# Bootloader selection
+#   Teensy       halfkay
+#   Pro Micro    caterina
+#   Atmel DFU    atmel-dfu
+#   LUFA DFU     lufa-dfu
+#   QMK DFU      qmk-dfu
+#   ATmega32A    bootloadHID
+#   ATmega328P   USBasp
+BOOTLOADER = atmel-dfu
+
+# Extra source files for IS3731 lighting
+SRC = TWIlib.c issi.c lighting.c
